requirements.txt
Django==3.2.20
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
djangorestframework==3.14.0
djoser==2.1.0
psycopg2==2.9.6
django-autoslug==1.9.9
django-filter==23.2
drf-extra-fields==3.6.1
Pillow==9.5.0